requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.17
supabase==2.10.0
anthropic==0.39.0
psutil==6.1.0
pydantic== 2.10.0